refactor: add components to lean the screen file
